Moved smooth change toggles into smooth mode
Moved the getCellChange check in the render loop only when it's in the smooth mode.

Makes other modes run faster.
Still having an issue when trying to increase the height scale of the pillar while in smooth mode.

The height does not change, but the size of the field does when changing the height for some reason