Changed how the wind picks up sand
Before it was if the sand is above the treshold it will always pickup and below the threshold never.

I turned it into a percent (if above threshold always, if below threshold that percent close to the treshold).

For best results set treshold to the same as your highest peak